Fixed Portal Bug
Fixed the bug that when you go backwards into a portal, you come out
the other as if you went forwards which will send you back into the
second portal and out the first etc. so you will zigzag between
the portals.
Fixed bug by adding bool 'IsGoingBackwards' so when you go
backwards, the portal will know and send you out the other with
a velocity of -40 instead of 40.
Also compiled code.